refactor wayland backend code
