+++ Cargo.toml
@@ -28,11 +28,12 @@ xbuild = "0.2.0"
 clap = "4.5.27"
 
 [target.'cfg(target_os = "android")'.dependencies]
+tracing = "0.1.41"
 jni = "0.21.1"
 mockall = "0.13.1"
 toml = "0.8.12"
 android_logger = { version = "0.14", default-features = false }
-winit = { version = "0.30.9", features = [
+winit = { version = "0.30.11", features = [
     "android-native-activity",
     "wayland",
 ] }
