+++ manifest.yaml
@@ -18,3 +18,6 @@ android:
       - name: android.permission.INTERNET
       - name: android.permission.ACCESS_NETWORK_STATE
       - name: android.permission.POST_NOTIFICATIONS
+    uses_feature:
+      - name: android.hardware.type.pc
+        required: false
