+++ src/lib.rs
@@ -3,16 +3,11 @@
 pub mod android_main;
 pub mod app {
     pub mod build;
-    pub mod event_centralizer;
-    pub mod event_handler;
     pub mod run;
-}
-pub mod wayland {
-    pub mod compositor;
-    pub mod element;
-    pub mod input;
-    pub mod keymap;
-    pub mod winit_backend;
+    pub mod backend {
+        pub mod wayland;
+        pub mod webview;
+    }
 }
 pub mod proot {
     pub mod launch;
